Move filament ID code to separate file
