+++ ps3mf/model.go
@@ -2,7 +2,6 @@ package ps3mf
 
 import (
 	"bufio"
-	"encoding/json"
 	"encoding/xml"
 	"github.com/hpinc/go3mf"
 	"github.com/hpinc/go3mf/importer/stl"
@@ -10,30 +9,6 @@ import (
 	"mosaicmfg.com/stl-to-3mf/util"
 	"os"
 )
-
-// FilamentIDMap contains a mappingfrom project input index to Element filamentId
-// (only used for Element processing)
-type FilamentIDMap map[byte]byte
-
-type filamentIDJsonData struct {
-	IDs [][]byte `json:"filamentIds"`
-}
-
-func UnmarshalFilamentIds(serialized string) (FilamentIDMap, error) {
-	var parsed filamentIDJsonData
-	if err := json.Unmarshal([]byte(serialized), &parsed); err != nil {
-		return nil, err
-	}
-
-	ids := make(FilamentIDMap)
-	for _, pair := range parsed.IDs {
-		key := pair[0]
-		value := pair[1]
-		ids[key] = value
-	}
-
-	return ids, nil
-}
 
 type ModelOpts struct {
 	Name           string
